+++ packages/contracts/CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Upcoming]
+
 ### Added
 
-- Added a `ERC165ContractRegistry` `Component` to register contracts by their address and ERC165 interface ID.
+- Added a `ERC165Registry` `Component` to register contracts by their address and ERC165 interface ID.
   
 ## v0.2.0-alpha